Export app functions, not instances
- it's now possible to later replace global config with parameterized
  config
- removes the need for the _TEST_ONLY_redisClient hack
