+++ apigw/src/enduser/__tests__/csrf.ts
@@ -3,9 +3,10 @@
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
 import { GatewayTester } from '../../shared/test/gateway-tester'
-import app from '../app'
 import { csrfCookieName } from '../../shared/middleware/csrf'
 import { CitizenUser } from '../../shared/service-client'
+import enduserGwApp from '../app'
+import { createRedisClient } from '../../shared/redis-client'
 
 const mockUser: CitizenUser = {
   id: '4f73e4f8-8759-46c6-9b9d-4da860138ce2'
@@ -14,6 +15,7 @@ const mockUser: CitizenUser = {
 describe('CSRF middleware and cookie handling in enduser-gw', () => {
   let tester: GatewayTester
   beforeAll(async () => {
+    const app = enduserGwApp(createRedisClient())
     tester = await GatewayTester.start(app, 'enduser')
   })
   beforeEach(async () => tester.login(mockUser))
