+++ apigw/src/shared/__tests__/saml-slo.ts
@@ -17,6 +17,7 @@ import { fromCallback } from '../promise-utils'
 import type { CitizenUser } from '../service-client'
 import { sessionCookie } from '../session'
 import { GatewayTester } from '../test/gateway-tester'
+import { createRedisClient } from '../redis-client'
 
 const mockUser: CitizenUser = {
   id: '942b9cab-210d-4d49-b4c9-65f26390eed3'
@@ -59,6 +60,7 @@ describe('SAML Single Logout', () => {
   const sfiMock = config.sfiMock
 
   beforeEach(async () => {
+    redisClient = createRedisClient()
     // In order to enable the REAL Suomi.fi passport-saml Strategy only for
     // these tests, config.sfiMock should be true in every other case but this
     // test suite + as Strategy vs. DummyStrategy selection is done at app
@@ -68,10 +70,8 @@ describe('SAML Single Logout', () => {
     // instead of beforeAll.
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     ;(config as any).sfiMock = false
-    const { default: app, _TEST_ONLY_redisClient } = await import(
-      '../../enduser/app'
-    )
-    redisClient = _TEST_ONLY_redisClient
+    const { default: enduserGwApp } = await import('../../enduser/app')
+    const app = enduserGwApp(redisClient)
     tester = await GatewayTester.start(app, 'enduser')
   })
   afterEach(async () => {
